% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{chunk}
\alias{chunk}
\title{Chunk a vector or list}
\usage{
chunk(x, len = NULL, n = NULL)
}
\arguments{
\item{x}{vector of list}

\item{len}{target length of chunks}

\item{n}{number of chunks}
}
\description{
\code{chunk} takes a vector (or list) and returns a list of chunks
of (approximately) equal to a specified length.
}
\details{
If \code{n} is specified, \code{len} is ignored and \code{chunk} returns
 a list of length \code{n} of "chunks" of \code{x}. Otherwise
 \code{n} is calculated to break the vector into chunks which are
 each approximately of length \code{len}. If both \code{len} and
 \code{n} are unspecified, \code{chunk} simply returns \code{x}.
}
\examples{
chunk(letters, 10)
chunk(LETTERS, n = 3)

}
