% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{corpuses}
\alias{corpuses}
\title{Google n-gram corpus information}
\format{a 22 x 4 ngram data frame}
\usage{
corpuses
}
\description{
Details of the various corpuses available through the Google n-gram took
}
\keyword{datasets}

