% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hacker}
\alias{hacker}
\title{Sample n-gram data}
\format{
a 236 x 4 ngram data frame
}
\usage{
hacker
}
\description{
Frequency data for the phrases "hacker", "programmer", from 1950 to 2008.
}
\keyword{datasets}
