\name{ngram-package}
\alias{ngram-package}
\alias{ngram}
\docType{package}
\title{
ngram - Google n-grams viewer
}
\description{
Query the Google Books Ngram Viewer data.
Given a list of phrases, obtain the frequencies of these phrases 
in a corpus of books (e.g., "British English", "English Fiction", "French") 
}
\details{
\tabular{ll}{
Package: \tab ngram\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-06\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
