% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngramr-package.R
\docType{package}
\name{ngramr}
\alias{ngramr}
\alias{ngramr-package}
\title{ngramr: Dig into the Google Ngram Viewer using R}
\description{

}
\references{
Michel, Jean-Baptiste, et al. "Quantitative analysis of culture using
millions of digitized books." \emph{Science} 331, No. 6014 (2011): 176--182.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seancarmody/ngramr}
  \item Report bugs at \url{https://github.com/seancarmody/ngramr/issues}
}

}
\author{
\strong{Maintainer}: Sean Carmody \email{seancarmody@gmail.com} [copyright holder]

}
\keyword{internal}
