% Generated by roxygen2 (4.0.1): do not edit by hand
\name{theme_google}
\alias{theme_google}
\title{Google Ngram theme for ggplot2}
\usage{
theme_google(...)
}
\arguments{
\item{...}{additional parameters to pass to \code{theme}}
}
\description{
Google Ngram theme for ggplot2
}
\details{
Use a Google Ngram-style plot theme.
}

