\name{ngrami}
\alias{ngrami}
\title{Get n-gram frequencies (case insensitive version)}
\usage{
  ngrami(phrases, aggregate = TRUE, ...)
}
\arguments{
  \item{phrases}{vector of phrases}

  \item{aggregate}{sum up each of the terms}

  \item{...}{remaining parameters passed to ngram}
}
\description{
  Get n-gram frequencies (case insensitive version)
}
\details{
  Possible corpora: eng_2012, eng_2009, eng_us_2012,
  eng_us_2009, eng_gb_2012, eng_gb_2009, chi_sim_2012,
  chi_sim_2009, fre_2012, fre_2009, ger_2012, ger_2009,
  spa_2012, spa_2009, rus_2012, rus_2009, heb_2012,
  heb_2009, ita_2012, eng_fiction_2012, eng_fiction_2009,
  eng_1m_2009
}

