% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrami.R
\name{ngrami}
\alias{ngrami}
\title{Get n-gram frequencies (case insensitive version)}
\usage{
ngrami(phrases, aggregate = TRUE, ...)
}
\arguments{
\item{phrases}{vector of phrases}

\item{aggregate}{sum up each of the terms}

\item{...}{remaining parameters passed to ngram}
}
\description{
Get n-gram frequencies (case insensitive version)
}
