% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ngram}
\alias{print.ngram}
\title{Print n-gram contents}
\usage{
\method{print}{ngram}(x, rows = 6, ...)
}
\arguments{
\item{x}{ngram object as returned by \code{link{ngram}}}

\item{rows}{number of rows to print from top and bottom of ngram data set. Default is 6.}

\item{...}{additional parameters passed to default print method.}
}
\description{
Print n-gram contents
}
\examples{
x <- ngram(c("hacker", "programmer"), year_start = 1950)
x
print(x, rows = 3)
}
